Refactor the Compiler and Resolver structs to not have the horrible linked list structure and be able to support Closures without rust freaking out
Replace it by using the functions Vec and the current_chunk (and
related) methods. This lets us not have to manage child compilers and
just switch which FunctionChunk we write to. No need to manage refs to
parents or shared refs to a Resolver or updating values afterwards.

Speaking of the resolver: Switched to using a similar idea with a
growing stack of ResolverNodes, again allowing for new ResolverNodes to
be pushed and modified at will without needing references to parents.