Finished* chapter 27
Asterisk because it's broken. LoxInstances are COPIED from the globals
vec, so modifications to it's data aren't kept. I have to add LoxPointers and a fake
"heap" to make this work properly. So then I also need to add garbage
collection. Sick...

I was so happy I was able to skip implementing the GC chapter because
rust was so kind as to take care of it all for me. But not anymore I
guess.