Remove the clone in get_variable_name
Turns out this is usually never required, because the returned string is
usually just used in a hashmap lookup anyway. So a &String suffices

The cases where clones are needed are then the values are actually being
inserted into the hashmap. (idea: can we abstract away the strings
entierly and just hash usizes, representing the index of the loxstring?
Or even better just make it a vector?)