WIP flatten the globals hashmap into a vec
Right now it's piggypacking off of the constants vector and creating a
bunch of extra Undef placeholders. Fix this by modifying the compiler
(either with a seperate vec for all loxstrings or resolving the globals)

Alternatively, write a resolver pass? This seems to get close to optimal
memory usage while also not requiring a second pass