Make the globals vec in VM slightly better
Make it only the size of every identifier constant, instead of every
constant. Did this by adding yet another vec to the Compiler which gets
used in identifier_constant(). Now each OpCode with an index to an
identifier indexes into an identifier specific vec, instead of just the
generic constants one.

This means we have two vecs where we had one, but since they serve distinct purposes I'm
fine with it